refactor(mobile): split thread_detail_page into part files
thread_detail_page.dart was 996 lines by the ratchet's count
(check-file-sizes-core.mjs splits on newlines, so a 995-line
newline-terminated file counts 996) against a hard limit of 1000.
Four lines of headroom left no room for feature work.

Moves _NestedThreadSummaryRow and _ThreadMessage/_Avatar into part
files under thread_detail_page/, matching the convention already used
by channels_page.dart and activity_page.dart. Parent drops to 603.

Pure relocation: reconstructing the original from the three files is
byte-identical to the pre-split content. No behavior change.
